template_matching/requirements.txt
numpy>1.20.3
opencv_python_headless==4.5.4.60
Pillow==9.2.0
streamlit==1.10.0
